test: split shared inference provider harness
